dry up by refactoring to a mixin
