+++ resources/js/components/blueprints/Builder.vue
@@ -50,9 +50,12 @@
 </template>
 
 <script>
+import SuggestsConditionalFields from './SuggestsConditionalFields';
 import Tabs from './Tabs.vue';
 
 export default {
+
+    mixins: [SuggestsConditionalFields],
 
     components: {
         Tabs,
@@ -64,12 +67,6 @@ export default {
         showTitle: Boolean,
         useTabs: { type: Boolean, default: true },
         isFormBlueprint: { type: Boolean, default: false },
-    },
-
-    provide() {
-        return {
-            suggestableConditionFieldsProvider: this.makeConditionsProvider(),
-        }
     },
 
     data() {
@@ -156,26 +153,6 @@ export default {
             this.errors = {};
             this.$dirty.remove('blueprints');
         },
-
-        makeConditionsProvider() {
-            const provide = {};
-            Object.defineProperties(provide, {
-                suggestableFields: { get: () => this.suggestableConditionFields },
-            });
-            return provide;
-        },
-
-        getFieldsFromImportedFieldset(fieldset, prefix) {
-            return Statamic.$config.get(`fieldsets.${fieldset}.fields`, [])
-                .reduce((fields, field) => {
-                    return fields.concat(
-                        field.type === 'import'
-                            ? this.getFieldsFromImportedFieldset(field.fieldset, field.prefix)
-                            : [field.handle]
-                    );
-                }, [])
-                .map(handle => prefix ? `${prefix}${handle}` : handle);
-        }
 
     }
